Fixed database
A database complete with schema and dump is now provided when building the  and publishing the project. The DBFacade class has been changed to rely on this self contained DB file, instead of the arbitrary, non-reliable /tmp/chirp.db.

Furthermore, the DBFacade now loads this file independent on the systems directory structure, allowing it to be loaded from an abitrary machine build or publish