Usertimeline with following cheeps
Changed the GetCheepsFromAuthorAndFollowers to find followers cheeps by giving the follower id to GetCheepsFromAuthor.
This was done to circumvent the problem of Lists lazy loading with EF-Core and Linq.
With the previous method of taking cheeps form the follower cheeps directly from the original author, the follower cheeps are not loaded into the program at all.