UserTimeline shows following and own cheeps correctly
Added a SignInManager to check whether or not the requesting user is authenticated, and whether or not the user is accessing their own timeline. If a user accesses their own timeline, both their on and the cheeps of their followers are displayed. If a user accesses another users timelines, only the other users cheeps are displayed.